pyproject.toml
[tool.poetry]
name = "just-practice"
version = "0.1.0"
description = ""
authors = ["razdva"]
readme = "README.md"
packages = [{include = "just_practice"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.1.0"
fastapi = {extras = ["all"], version = "^0.108.0"}
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
